use 27SF512 for char ROM, GAL22V10 for shift reg
The 27SF512 is the fastest programmable ROM I have,
at 70 ns, so in theory could produce valid character
pixel row data after only two clock cycles.
It's a bit less "authentic" than, let's say, a
27C256-70, but I don't have any of those, and
the T48 should be able to program the 27SF512.

Originally I was going to use a '166 shift register
for the pixel output, but the 74ACT166 is virtually
nonexistent. Apparently Hitachi made them (there's a
datasheet), but I can't find any to buy. A GAL22V10
can easily be a shift register, so no worries there.