more work on VRAM host interfacing
Fixed error in IDT7134 symbol.

Working on address decoding and lower/upper VRAM
selection. Since we're using a GAL to decode the host
control signals, it can just do its own logic to
recognize when the upper part of the I/O region
is being accessed. The CPU/glue logic module generates
-IOR4..-IOR7 signals when this region is accessed,
but they're not present on the bus. However, the
control signal decoding GAL can just look at
A8..A15 to figure out when the VRAM is being
accessed.