reverted previous VCount change; hw was correct
The TC a.k.a. "RCO" output only goes high if the
counter is at terminal count, AND the CET input
is high. Since the low-bit '163 in VCount has its
CET driven by HCOUNT_END, it only asserts TC/RCO
if HCOUNT_END is asserted, thus the cascaded counters
are never enabled if the low counter is not enabled.

This is good news since it means I don't need to
change the wiring of the prototype VCount module.

The change being reverted was not really wrong.
(I don't think it was wrong anyway.) However, I
think it's important for the schematics to be
100% consistent with the prototype hardware.