limit memfetch+raster by VActive signal
I think I have the timing of the VActive signal
somewhat correct: it starts at the correct time,
when VCOUNT_END is asserted, i.e., just before
the last (non-visible) scanline of the frame.
The ReadoutCtrl module does appear to start generating
the appropriate signals just before the end of the
last non-visible scanline (in preparation for
the first visible scanlinen.)

I'm not sure whether VActive gets turned off at
the correct time yet. Need to think about this
(and, of course, run the simulation to see what
happens.)