working on Readout and ReadoutCtrl implementation
Have a preliminary Readout circuit with '163 (readout
address) and '273 (row begin address) chips.

Added inputs and outputs to ReadoutCtrl module,
which will be a GAL. (Hopefully everything will fit
in a single 22V10.) This module will "run the show"
regarding generation of readout addresses and control
signals for the registers storing the character and
attrtibute data.