begin horiontal activity two cycles later
Also, start the ReadoutCtrl two-bit counter at 2
rather than 0.

These changes should allow the initial character to
be latched without incrementing the readout counter.
Assuming that the readout address increment will
be initiated when the counter is 0, then as
originally implemented, the readout address would
be 1 rather than 0 when the first character byte
is fetched from VRAM, which is not what we want.

The new scheme should allow for even-numbered VRAM
addresses to be characters and odd-numbered VRAM
addresses to be attributes.